+++ CHANGES.md
@@ -5,6 +5,10 @@
 ### Changed
 
 - Updated stac-fastapi libraries to v2.5.3 ([#101](https://github.com/stac-utils/stac-fastapi-pgstac/pull/101))
+- 
+### Added
+
+- Ability to configure the database runtime parameters ([#92](https://github.com/stac-utils/stac-fastapi-pgstac/pull/92))
 
 ## [2.4.11] - 2023-12-01